renamed package and small tweak to BoardElementImageFinder
